Moved MidiOut into midi package
