+++ src/main/java/net/perkowitz/issho/controller/MidiSetup.java
@@ -10,6 +10,7 @@ import lombok.extern.java.Log;
 import net.perkowitz.issho.controller.midi.Clockable;
 import net.perkowitz.issho.controller.midi.DeviceRegistry;
 import net.perkowitz.issho.controller.midi.MidiIn;
+import net.perkowitz.issho.controller.midi.MidiOut;
 import net.perkowitz.issho.controller.novation.LaunchpadPro;
 import net.perkowitz.issho.controller.yaeltex.YaeltexHachiXL;
