+++ src/main/java/net/perkowitz/issho/controller/midi/DeviceRegistry.java
@@ -2,11 +2,8 @@ package net.perkowitz.issho.controller.midi;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import com.google.common.collect.Sets;
 import lombok.Setter;
 import lombok.extern.java.Log;
-import net.perkowitz.issho.controller.Controller;
-import net.perkowitz.issho.controller.MidiOut;
 import net.perkowitz.issho.controller.novation.LaunchpadPro;
 import net.perkowitz.issho.controller.yaeltex.YaeltexHachiXL;
