+++ src/main/java/net/perkowitz/issho/controller/yaeltex/Test.java
@@ -3,10 +3,8 @@ package net.perkowitz.issho.controller.yaeltex;
 
 import net.perkowitz.issho.controller.Colors;
 import net.perkowitz.issho.controller.Controller;
-import net.perkowitz.issho.controller.MidiOut;
 import net.perkowitz.issho.controller.MidiSetup;
 import net.perkowitz.issho.controller.elements.Button;
-import net.perkowitz.issho.controller.elements.Knob;
 import net.perkowitz.issho.controller.elements.Pad;
 
 import java.awt.*;
